use mixin to have get/set values for node list
