+++ pyproject.toml
@@ -42,7 +42,6 @@ dev = [
     "pytest>=8.0,<9",
     "pytest-cov>=4.1,<5",
     "pytest-mock>=3.12,<4",
-    "pytest-forked",
     "pre-commit>=3.6,<4",
     "graphviz>=0.17,<1",
     "h5py>=3.11,<4",
